fix: build error fix (강제로 마운트 뒤에 실행)
