+++ apps/web/src/app/dev/local/page.tsx
@@ -1,33 +1,14 @@
-'use client';
+import dynamic from 'next/dynamic';
 
-import { useEffect } from 'react';
-import { useRouter, useSearchParams } from 'next/navigation';
+const LazyLoginProgress = dynamic(() => import('./LoginProgress'), { ssr: false });
 
-import { useLogin } from '@/store/user';
-
-function DevLocalLoginPage() {
-  const searchParams = useSearchParams();
-  const token = searchParams.get('token');
-
-  const router = useRouter();
-  const { login } = useLogin();
-
-  const onLogin = async (token: string) => {
-    try {
-      await login(token);
-      router.replace('/mypage');
-    } catch (error) {
-      router.replace('/');
-    }
-  };
-
-  useEffect(() => {
-    if (token) {
-      onLogin(token);
-    }
-  }, [token]);
-
-  return <div>Loading...</div>;
+function DevLocalLoginPage({ searchParams }: { searchParams: URLSearchParams }) {
+  console.log('searchParams: ', searchParams);
+  return (
+    <>
+      <LazyLoginProgress />
+    </>
+  );
 }
 
 export default DevLocalLoginPage;
